pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "designing-real-world-ai-agents-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    {name = "Paul Iusztin", email = "p.b.iusztin@gmail.com"},
]
requires-python = ">=3.14"
dependencies = [
    "click>=8.3.1",
    "fastmcp>=2.0.0",
    "google-genai>=1.68.0",
    "opik>=1.10.52",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/research", "src/writing"]
